move more classes to sft/utils.py
